FIXED: Few compile warnings (conversions, unitialized vars, obsoletes, unreachable code etc.) FIXED: Improper memory deallocation in ~Spell(), thx Kenjiro, http://arcemu.org/forums/index.php?showtopic=14502&hl= FIXED: Mem. leak in InstanceMgr class

+++ src/arcemu-world/TaxiHandler.cpp
@@ -102,7 +102,7 @@ void WorldSession::SendTaxiList(Creature* pCreature)
 	TaxiMask[field] |= 1 << ((curloc-1)%32);
 
 	//Remove nodes unknown to player
-	for(int i = 0; i < 12; i++)
+	for(uint8 i = 0; i < 12; i++)
 	{
 		TaxiMask[i] &= GetPlayer( )->GetTaximask(i);
 	}
